Add comments and organize codebase
