+++ thirdeye/thirdeye-pinot/src/main/java/com/linkedin/thirdeye/anomalydetection/performanceEvaluation/AnomalyPercentagePerformanceEvaluation.java
@@ -1,11 +1,17 @@
 package com.linkedin.thirdeye.anomalydetection.performanceEvaluation;
 
 import com.linkedin.thirdeye.datalayer.dto.MergedAnomalyResultDTO;
+import com.linkedin.thirdeye.util.IntervalUtils;
 import java.util.List;
 import org.joda.time.Interval;
 
 
-public class AnomalyPercentagePerformanceEvaluation implements PerformanceEvaluate{
+/**
+ * The performance evaluator calculates the percentage of time when there are anomalies
+ * We care about the frequency of anomaly generated by our algorithm. This performance evaluator can help us providing
+ * this information to clients or backend algorithms
+ */
+public class AnomalyPercentagePerformanceEvaluation extends BasePerformanceEvaluate{
   private Interval windowInterval;
   private List<MergedAnomalyResultDTO> detectedResults;
 
@@ -19,8 +25,8 @@ public class AnomalyPercentagePerformanceEvaluation implements PerformanceEvalua
     long anomalyLength = 0;
     long totalLength = windowInterval.toDurationMillis();
 
-    List<Interval> anomalyIntervals = MergedAnomalyIntervalUtils.mergedAnomalyResultsToIntervals(detectedResults);
-    anomalyIntervals = MergedAnomalyIntervalUtils.mergeIntervals(anomalyIntervals);
+    List<Interval> anomalyIntervals = mergedAnomalyResultsToIntervals(detectedResults);
+    anomalyIntervals = IntervalUtils.mergeIntervals(anomalyIntervals);
 
     for(Interval interval : anomalyIntervals) {
       anomalyLength += interval.toDurationMillis();
