+++ thirdeye/thirdeye-pinot/src/main/java/com/linkedin/thirdeye/anomalydetection/performanceEvaluation/F1ScoreByTimePerformanceEvaluation.java
@@ -8,7 +8,7 @@ import java.util.List;
  * The f1 score of the cloned function with regarding to the labeled anomalies in original function.
  * The calculation is based on the time overlapped with the labeled anomalies.
  */
-public class F1ScoreByTimePerformanceEvaluation implements PerformanceEvaluate {
+public class F1ScoreByTimePerformanceEvaluation extends BasePerformanceEvaluate {
   private List<MergedAnomalyResultDTO> knownAnomalies;      // The merged anomalies which are labeled by user
   private List<MergedAnomalyResultDTO> detectedAnomalies;        // The merged anomalies which are generated by anomaly function
