+++ thirdeye/thirdeye-pinot/src/main/java/com/linkedin/thirdeye/anomalydetection/performanceEvaluation/PrecisionByTimePerformanceEvaluation.java
@@ -2,6 +2,7 @@ package com.linkedin.thirdeye.anomalydetection.performanceEvaluation;
 
 import com.linkedin.thirdeye.api.DimensionMap;
 import com.linkedin.thirdeye.datalayer.dto.MergedAnomalyResultDTO;
+import com.linkedin.thirdeye.util.IntervalUtils;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -13,12 +14,12 @@ import org.joda.time.Interval;
  * The calculation is based on the time overlapped with the labeled anomalies.
  * precision = (the overlapped time duration between detected anomalies and labelled anomalies) / (the time length of the detected anomalies)
  */
-public class PrecisionByTimePerformanceEvaluation implements PerformanceEvaluate {
+public class PrecisionByTimePerformanceEvaluation extends BasePerformanceEvaluate {
   private Map<DimensionMap, List<Interval>> knownAnomalyIntervals;      // The merged anomaly intervals which are labeled by user
   private List<MergedAnomalyResultDTO> detectedAnomalies;        // The merged anomalies which are generated by anomaly function
 
   public PrecisionByTimePerformanceEvaluation(List<MergedAnomalyResultDTO> knownAnomalies, List<MergedAnomalyResultDTO> detectedAnomalies) {
-    this.knownAnomalyIntervals = MergedAnomalyIntervalUtils.mergedAnomalyResultsToIntervalMap(knownAnomalies);
+    this.knownAnomalyIntervals = mergedAnomalyResultsToIntervalMap(knownAnomalies);
     this.detectedAnomalies = detectedAnomalies;
   }
 
@@ -27,8 +28,8 @@ public class PrecisionByTimePerformanceEvaluation implements PerformanceEvaluate
     if(knownAnomalyIntervals == null || knownAnomalyIntervals.size() == 0) {
       return 0;
     }
-    Map<DimensionMap, List<Interval>> anomalyIntervals = MergedAnomalyIntervalUtils.mergedAnomalyResultsToIntervalMap(detectedAnomalies);
-    MergedAnomalyIntervalUtils.mergeIntervals(anomalyIntervals);
+    Map<DimensionMap, List<Interval>> anomalyIntervals = mergedAnomalyResultsToIntervalMap(detectedAnomalies);
+    IntervalUtils.mergeIntervals(anomalyIntervals);
     Map<DimensionMap, Long> dimensionToDetectedAnomalyTimeLength = new HashMap<>();
     Map<DimensionMap, Long> dimensionToOverlapTimeLength = new HashMap<>();
