+++ thirdeye/thirdeye-pinot/src/main/java/com/linkedin/thirdeye/anomalydetection/performanceEvaluation/RecallByTimePreformanceEvaluation.java
@@ -2,6 +2,7 @@ package com.linkedin.thirdeye.anomalydetection.performanceEvaluation;
 
 import com.linkedin.thirdeye.api.DimensionMap;
 import com.linkedin.thirdeye.datalayer.dto.MergedAnomalyResultDTO;
+import com.linkedin.thirdeye.util.IntervalUtils;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -13,13 +14,13 @@ import org.joda.time.Interval;
  * The calculation is based on the time overlapped with the labeled anomalies.
  * recall = (the overlapped time duration between detected anomalies and labelled anomalies) / (the time length of the labelled anomalies)
  */
-public class RecallByTimePreformanceEvaluation implements PerformanceEvaluate {
+public class RecallByTimePreformanceEvaluation extends BasePerformanceEvaluate {
   private Map<DimensionMap, List<Interval>> knownAnomalyIntervals;      // The merged anomaly intervals which are labeled by user
   private List<MergedAnomalyResultDTO> detectedAnomalies;        // The merged anomalies which are generated by anomaly function
   private Map<DimensionMap, Long> knownAnomalyTimeLength;
 
   public RecallByTimePreformanceEvaluation(List<MergedAnomalyResultDTO> knownAnomalies, List<MergedAnomalyResultDTO> detectedAnomalies) {
-    this.knownAnomalyIntervals = MergedAnomalyIntervalUtils.mergedAnomalyResultsToIntervalMap(knownAnomalies);
+    this.knownAnomalyIntervals = mergedAnomalyResultsToIntervalMap(knownAnomalies);
     this.knownAnomalyTimeLength = new HashMap<DimensionMap, Long>();
     for(Map.Entry<DimensionMap, List<Interval>> entry : this.knownAnomalyIntervals.entrySet()) {
       long timeLength = 0l;
@@ -36,8 +37,8 @@ public class RecallByTimePreformanceEvaluation implements PerformanceEvaluate {
     if(knownAnomalyIntervals == null || knownAnomalyIntervals.size() == 0) {
       return 0;
     }
-    Map<DimensionMap, List<Interval>> anomalyIntervals = MergedAnomalyIntervalUtils.mergedAnomalyResultsToIntervalMap(detectedAnomalies);
-    MergedAnomalyIntervalUtils.mergeIntervals(anomalyIntervals);
+    Map<DimensionMap, List<Interval>> anomalyIntervals = mergedAnomalyResultsToIntervalMap(detectedAnomalies);
+    IntervalUtils.mergeIntervals(anomalyIntervals);
     Map<DimensionMap, Long> dimensionToOverlapTimeLength = new HashMap<>();
 
     for(MergedAnomalyResultDTO detectedAnomaly : detectedAnomalies) {
