+++ thirdeye/thirdeye-pinot/src/main/java/com/linkedin/thirdeye/util/IntervalUtils.java
@@ -1,4 +1,4 @@
-package com.linkedin.thirdeye.anomalydetection.performanceEvaluation;
+package com.linkedin.thirdeye.util;
 
 import com.linkedin.thirdeye.api.DimensionMap;
 import com.linkedin.thirdeye.datalayer.dto.MergedAnomalyResultDTO;
@@ -12,7 +12,7 @@ import java.util.Stack;
 import org.joda.time.Interval;
 
 
-public class MergedAnomalyIntervalUtils {
+public class IntervalUtils {
   /**
    * This method is designed to merge a list of intervals to a list of intervals with no overlap in between
    * @param intervals
@@ -64,35 +64,5 @@ public class MergedAnomalyIntervalUtils {
     for(DimensionMap dimension : anomalyIntervals.keySet()) {
       anomalyIntervals.put(dimension ,mergeIntervals(anomalyIntervals.get(dimension)));
     }
-  }
-
-  /**
-   * convert merge anomalies to a dimension-intervals map
-   * @param mergedAnomalyResultDTOList
-   * @return
-   */
-  public static Map<DimensionMap, List<Interval>> mergedAnomalyResultsToIntervalMap (List<MergedAnomalyResultDTO> mergedAnomalyResultDTOList) {
-    Map<DimensionMap, List<Interval>> anomalyIntervals = new HashMap<>();
-    for(MergedAnomalyResultDTO mergedAnomaly : mergedAnomalyResultDTOList) {
-      if(!anomalyIntervals.containsKey(mergedAnomaly.getDimensions())) {
-        anomalyIntervals.put(mergedAnomaly.getDimensions(), new ArrayList<Interval>());
-      }
-      anomalyIntervals.get(mergedAnomaly.getDimensions()).add(
-          new Interval(mergedAnomaly.getStartTime(), mergedAnomaly.getEndTime()));
-    }
-    return anomalyIntervals;
-  }
-
-  /**
-   * convert merge anomalies to interval list without considering the dimension
-   * @param mergedAnomalyResultDTOList
-   * @return
-   */
-  public static List<Interval> mergedAnomalyResultsToIntervals (List<MergedAnomalyResultDTO> mergedAnomalyResultDTOList) {
-    List<Interval> anomalyIntervals = new ArrayList<>();
-    for(MergedAnomalyResultDTO mergedAnomaly : mergedAnomalyResultDTOList) {
-      anomalyIntervals.add(new Interval(mergedAnomaly.getStartTime(), mergedAnomaly.getEndTime()));
-    }
-    return anomalyIntervals;
   }
 }
